Prometheus/Loki: Improve operations header UX (#46346)
* Prometheus/Loki: Improve operations header UX

* More tweaks

* Rename file to match compponent
