+++ public/app/plugins/datasource/prometheus/querybuilder/shared/OperationEditor.tsx
@@ -1,6 +1,6 @@
 import { css } from '@emotion/css';
 import { DataSourceApi, GrafanaTheme2 } from '@grafana/data';
-import { FlexItem, Stack } from '@grafana/experimental';
+import { Stack } from '@grafana/experimental';
 import { Button, useStyles2 } from '@grafana/ui';
 import React from 'react';
 import { Draggable } from 'react-beautiful-dnd';
@@ -11,8 +11,7 @@ import {
   QueryBuilderOperationDef,
   QueryBuilderOperationParamDef,
 } from '../shared/types';
-import { OperationInfoButton } from './OperationInfoButton';
-import { OperationName } from './OperationName';
+import { OperationHeader } from './OperationHeader';
 import { getOperationParamEditor } from './OperationParamEditor';
 import { getOperationParamId } from './operationUtils';
 
@@ -121,27 +120,15 @@ export function OperationEditor({
           {...provided.draggableProps}
           data-testid={`operations.${index}.wrapper`}
         >
-          <div className={styles.header} {...provided.dragHandleProps}>
-            <OperationName
-              operation={operation}
-              def={def}
-              index={index}
-              onChange={onChange}
-              queryModeller={queryModeller}
-            />
-            <FlexItem grow={1} />
-            <div className={`${styles.operationHeaderButtons} operation-header-show-on-hover`}>
-              <OperationInfoButton def={def} operation={operation} />
-              <Button
-                icon="times"
-                size="sm"
-                onClick={() => onRemove(index)}
-                fill="text"
-                variant="secondary"
-                title="Remove operation"
-              />
-            </div>
-          </div>
+          <OperationHeader
+            operation={operation}
+            dragHandleProps={provided.dragHandleProps}
+            def={def}
+            index={index}
+            onChange={onChange}
+            onRemove={onRemove}
+            queryModeller={queryModeller}
+          />
           <div className={styles.body}>{operationElements}</div>
           {restParam}
           {index < query.operations.length - 1 && (
@@ -205,16 +192,6 @@ const getStyles = (theme: GrafanaTheme2) => {
       marginBottom: theme.spacing(1),
       position: 'relative',
     }),
-    header: css({
-      borderBottom: `1px solid ${theme.colors.border.medium}`,
-      padding: theme.spacing(0.5, 0.5, 0.5, 1),
-      gap: theme.spacing(1),
-      display: 'flex',
-      alignItems: 'center',
-      '&:hover .operation-header-show-on-hover': css({
-        opacity: 1,
-      }),
-    }),
     infoIcon: css({
       color: theme.colors.text.secondary,
     }),
@@ -233,12 +210,6 @@ const getStyles = (theme: GrafanaTheme2) => {
       fontWeight: theme.typography.fontWeightMedium,
       verticalAlign: 'middle',
       height: '32px',
-    }),
-    operationHeaderButtons: css({
-      opacity: 0,
-      transition: theme.transitions.create(['opacity'], {
-        duration: theme.transitions.duration.short,
-      }),
     }),
     paramValue: css({
       display: 'table-cell',
